Deleted a misleading comment.
We do not convert the freq from 16~17 Hz to 10 and then 10 to 30. The comment was for the previous approach where we used only 10 Hz frequency. The function, " convert_freq" uses the "signal" package to increase frequency from 16~17 to 30HZ. We used the floor function just to calculate the frequency for each second. Ethica data's frequency differs from second to second. It is either 16 or 17 HZ. To use the signal package we needed the initial frequency( to be more precise the initial period = 1/ freq), as well as final frequency.